+++ ebcpy/__init__.py
@@ -8,4 +8,4 @@ from .simulationapi.fmu import FMU_API
 from .optimization import Optimizer
 
 
-__version__ = '0.3.13'
+__version__ = '0.3.14'