+++ ebcpy/utils/__init__.py
@@ -4,10 +4,12 @@ Contains a statistics analyzer and a visualizer.
 """
 import logging
 import os
+from pathlib import Path
+from typing import Union
 
 
 def setup_logger(name: str,
-                 cd: str = None,
+                 working_directory: Union[Path, str] = None,
                  level=logging.DEBUG):
     """
     Setup an class or module specific logger instance
@@ -15,7 +17,7 @@ def setup_logger(name: str,
 
     :param str name:
         The name of the logger instance
-    :param str cd:
+    :param str,Path working_directory:
         The path where to store the logfile.
         If None is given, logs are not stored.
     :param str level:
@@ -35,9 +37,9 @@ def setup_logger(name: str,
     console = logging.StreamHandler()
     console.setFormatter(fmt=formatter)
     logger.addHandler(hdlr=console)
-    if cd is not None:
-        os.makedirs(cd, exist_ok=True)
-        file_handler = logging.FileHandler(filename=os.path.join(cd, f"{name}.log"))
+    if working_directory is not None:
+        os.makedirs(working_directory, exist_ok=True)
+        file_handler = logging.FileHandler(filename=working_directory.joinpath(f"{name}.log"))
         file_handler.setFormatter(fmt=formatter)
         logger.addHandler(hdlr=file_handler)
     return logger